Fixed Bug with the use of lastentry.py by hanime.py
hanime.py would change the last entry for hanime to the latest entry that it
gets, which is expected behaviour HOWEVER further iterations of the loop use
the new variable instead of the original one. Therefore the use of
``ha_before`` was used instead of ``ha`` to do the checking.